backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
tensorflow-macos==2.16.1
numpy==1.26.4
Pillow==10.1.0
scikit-learn==1.3.2
opencv-python==4.8.1.78
reportlab==4.0.7
python-multipart==0.0.6
werkzeug==3.0.1
matplotlib==3.8.2
requests==2.31.0
google-generativeai==0.8.3
imbalanced-learn==0.12.0
python-dotenv==1.0.0
